LPD-87314 Remove applicationsMenu.spec.ts
